pyproject.toml
[project]
name = "banana-slides"
version = "0.3.0"
description = "Banana Slides – AI native PPT generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.52.0",
    "openai>=1.0.0",
    "pydantic>=2.9.0",
    "pillow>=12.0.0",
    "python-pptx>=1.0.0",
    "python-dotenv>=1.0.1",
    "reportlab>=4.1.0",
    "werkzeug>=3.0.1",
    "markitdown[all]",
    "tenacity>=9.0.0",
    "alembic>=1.13.0",
    "flask-migrate>=4.0.0",
    "img2pdf>=0.5.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "flake8>=6.1.0",
    "black>=23.0.0",
]

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
